refactor: extract state machine to worker_state.go
